refacto + pass proper flag to install
